Change PHPDoc "description" fields to "summary
See http://phpdoc.org/docs/latest/guides/docblocks.html

Summary
    A short piece of text, usually one line, providing the basic
function of the associated element.
Description
    An optional longer piece of text providing more details on the
associated element’s function. This is very useful when working with a
complex element.
A series of tags
    These provide additional information in a structured manner. With
these tags you can link to other elements, provide type information for
properties and arguments, and more.

Summary

The summary is a short but effective overview of an element; you can
compare it to a slogan or headline. Summaries are plain and simple, they
cannot contain formatting or inline tags (see the next chapter for more
information on this) and can be used to skim through, for example, a
list of methods.

The summary can be separated from the description in two ways:

    With an empty whiteline:

    /**
     * This is a summary
     *
     * This is a description
     */

    Or by adding a period followed by a new-line:

    /**
     * This is a summary.
     * This is a description
     */

There is not much more to summaries, they are simple and straightforward
to use.